refactor: Create combinatorics.h file
Move the Combinatorics class definition to that file.
Include the header in calc_permutations_combinations.h
